refactor: implement Service layer for Skinny Controllers pattern
